LightPatchをShipStatusに移動
